Move run_loop function to request_handler module
