+++ bot/__main__.py
@@ -1,3 +1,3 @@
-from .request_loop import run_loop
+from .request_handler import run_loop
 
 run_loop()
